BasscBackend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.0
boto3>=1.28
python-dotenv>=1.0
gunicorn>=21.0
dj-database-url>=2.2.0
# MySQL 可选：pip install mysqlclient 或 PyMySQL
